refactor(core): extract intake configuration tests
